3.1.6

Conflicts:
	debian/mongodb-enterprise-server.docs
	debian/mongodb-enterprise-unstable-server.docs
	debian/mongodb-org-server.docs
	debian/mongodb-org-unstable-server.docs
	rpm/mongodb-enterprise-unstable.spec
	rpm/mongodb-enterprise.spec
	rpm/mongodb-org-unstable.spec
	rpm/mongodb-org.spec
	src/mongo/db/storage/wiredtiger/wiredtiger_record_store.cpp

+++ src/mongo/db/curop.h
@@ -156,8 +156,8 @@ public:
 
     // detailed options
     long long cursorid;
-    int ntoreturn;
-    int ntoskip;
+    long long ntoreturn;
+    long long ntoskip;
     bool exhaust;
 
     // debugging/profile info
@@ -187,7 +187,7 @@ public:
 
     // response info
     int executionTime;
-    int nreturned;
+    long long nreturned;
     int responseLength;
 };
 
@@ -324,7 +324,8 @@ public:
     uint64_t getRemainingMaxTimeMicros() const;
 
     //
-    // Methods for getting/setting elapsed time.
+    // Methods for getting/setting elapsed time. Note that the observed elapsed time may be
+    // negative, if the system time has been reset during the course of this operation.
     //
 
     void ensureStarted();
